Many changes
Now GraphicElements are not heierarchies, but simply there to manipulate
verticies. SpriteElements will handle the GUI elements. Sprites and text
will only draw in front of all others. If a background is needed,
recommend you use a big textured Quad. GUIManager now in charge of
displaying the little "hello there" string. Used to be that VRArray
would delete and re-create all elements every frame, but now it won't.
Adding elements will return a handle to that element, so the managers
can manage the elements themselves.