Settling on drawing method and cleaning up
Removed some unused things that weren't really going to be used, and
settled on a way of drawing things. The wierd thing was there was
world.draw, menu.draw, then there was graphics.draw. So what did
graphics.draw do when world and menu were already drawing?

The idea is that world and menu update a list of "things to draw" that
graphics contains. Although the list is now just references. The world
objects and menu objects contain the information of the things that they
draw themselves.