Quick and dirty World functionality
Added some structure to world, giving it entities, and giving levels the
ability to spawn entities, and giving entityManager a few things to do.
Currently EntityManager is using lists instead of a recycleArray. This
is only temporary. It's a quick and dirty solution. Also levels "spawn"
entities by creating new ones. I want to change it later to be Levels
have certain entity "master copies" and "spawning" is done by copying
these "master copies" into a recycleArray of entities.