Changed World, Entity, organizational structure
Improved the previous "Quick and Dirty" implementation to a more
permanent one. Does the same thing, but the new structure is better.
Entities are no longer an extendable super class for other types of
entities, instead, entities are all the exact same size, and only
reference "States", which are different for each type of entity. States
are what can be inherited, extended, and are hierarchical. Entities
currently use a simple spriteelement for appearances. Perhaps this is
all that is needed, being that you can edit any property of an entity
from its controlling State object, including it's spriteelement. Also,
Level added, and a test level that spawns test entities. Levels keep
track of a small number of "Reference" entities, and shows the reference
entities to the entity manager. The entity manager then creates copies
of those entities and adds them to it's entity list.