Added camera, separate sprite lists
There are now two sprite lists in graphicsManager, world sprites and gui
sprites. The gui sprites are for rendering menu stuff. The coordinates
translate directly to screen coords. While world sprites are drawn
relative to the camera position, and coords for those sprites translate
to where it is relative to camera's rotation, position, and scale.