Made some changes, undid some changes
Was going to create a Spawner system, so that levels wouldn't have to
have "entities" that they would present to the entMan, and only have
them copied, but decided it would be too complicated. I might do that
later. Also, created a system to draw and manage multiple sprites from
one entity. Entities now link to a single sprite, which is now
maintained in a spriteList in the EntityManager, and those sprites can
in turn link to other sprites in the spriteList.  Sprites are chained in
this way, and can be managed through the new State method
PositionSprites.