Created preliminary system for level structure.
Levels are built out of "blocks":basic building blocks. It imitates the
spawning style of entities. But it will have no state, and will not use
a RecycleArray. It will simply be a list, and the level itself will
store the blocks. Blocks will not be managed by the level however, as an
external BlockManager will do that for them. There will soon be many
different types of blocks, all extending "Block". But right now it's
just the one kind.