Re-added polygon graphics
Put a basiceffect primitive drawer in a new manager called
"PolyManager". You can use it to draw polygons for your state. Found it
was rather necessary to have this in order to have polygon based
collisions. It's better now that it's confined to a separate class
instead of spread all over graphicsManager.