Upgraded pose editor
Made pose editor a little more usable. Strange things happened when I
used the new function "graphMan.DrawString()". It was somehow writing
over a worldSprite when I used GetEmptyElement() to find an empty
spriteElement. Things meant to affect the string being drawn ended up
also changing the color and scale of other sprites in the world.
To fix this I changed the DrawStrings function to always create a new
SpriteElement and add it to worldSprites. Those get cleared each loop
anyway.
I realized that the cause of this was rooted in the fact that
worldSprites was in setDataMode of the recycleArray. This meant that it
was easy for anyone who had access to worldSprites to change the values
of any other spriteElement in existence.
I changed the worldSprites and GUISprites to both be in copy data mode.
This seems to make more sense anyway, although it comes at a cost of
doubling the number of SpriteElements floating around.